Jarvis/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
openai==0.27.0
python-dotenv==1.0.0
requests==2.28.2